Propagate kernel statistics from NVPTX compilation to XLA Executable.
This will allow the autotuner or any other the caller to make the decision on whether to accept the compilation based on how many register bytes were spilled rather than a binary choice of cancelling on register spilling. We can then replace cancel_if_reg_spill by using this information.

PiperOrigin-RevId: 837503360