[XLA] Avoid duplicated include paths in XLA build
The change that introduces googletest wrapper does not work reliably:
depending on the order of deps in BUILD file, googletest include paths
may end up before the wrapper ones. When this happens, the
ASSERT_OK/EXPECT_OK macros are not available.

Instead of relying on the include path order to make include_next work,
rename the upstream googletest files, and have the wrapper include
renamed paths.

PiperOrigin-RevId: 839773153